tryning to fix drift

+++ dyn_planner/path_searching/src/common/search_space.h
@@ -86,15 +86,13 @@ namespace hagen {
                 void insert_obstacles(std::vector<Rect> obstacles);
                 void insert_trajectory(std::vector<Rect> trajectory);
                 void search_all_obstacles();
-                // bool obstacle_free(Rect search_rect);
                 void setEnvironment(const dyn_planner::EDTEnvironment::Ptr& env);
-                // bool obstacle_free(Eigen::Vector3d search_rect);
                 bool obstacle_free(Eigen::Vector3d search_rect, double optimal_time);
                 Eigen::Vector3d sample_free();
                 Eigen::Vector3d sample();
                 std::vector<double> linspace(double start_in, double end_in, double step_size);
-                bool collision_free(Eigen::Vector3d start, Eigen::Vector3d end, int r);
-                bool collision_free(Eigen::Vector3d start, Eigen::Vector3d end, int r, double optimal_time);
+                bool collision_free(Eigen::Vector3d start, Eigen::Vector3d end, double r);
+                bool collision_free(Eigen::Vector3d start, Eigen::Vector3d end, double r, double optimal_time);
                 void insert_obstacle(Eigen::Vector3d index);
                 std::vector<Eigen::Vector3d> nearest_obstacles(Eigen::Vector3d x
                                     , int max_neighbours);
