+++ dyn_planner/path_searching/src/rrt_star/rrtbase.h
@@ -56,7 +56,7 @@ namespace hagen {
             int sample_taken;
             int max_samples;
             std::vector<Eigen::Vector2d> lengths_of_edges;
-            int resolution;
+            double resolution;
             double pro;
             double voxel_side_length;
             Eigen::Vector3d init_min_point;
@@ -101,7 +101,7 @@ namespace hagen {
                 int sample_taken;
                 int max_samples;
                 std::vector<Eigen::Vector2d> Q;
-                int r;
+                double r;
                 double prc;
                 double voxel_side_length;
                 Eigen::Vector3d init_min_point;
